example folder created
