+++ .github/workflows/ci.yml
@@ -54,7 +54,7 @@ jobs:
           ar rcs build/libmusig2.a build/musig2.o
           ranlib build/libmusig2.a
           clang -x c++ -std=c++11 tests/testmusig2.c -g -o test_run/test -L./build -lmusig2 -lsecp256k1 -lstdc++ -lgtest -lgtest_main -pthread
-          clang examplemusig2.c -g -o example_run/example -L./build -lmusig2 -lsecp256k1 
+          clang examples/examplemusig2.c -g -o example_run/example -L./build -lmusig2 -lsecp256k1 
 
       - name: Run gtests
         run: test_run/test
