+++ examples/config.h
@@ -1,0 +1,13 @@
+
+#define NR_SIGNERS       5 // Number of signers
+
+#define NR_MESSAGES      (const int)2    // Number of messages
+
+
+#define MSG_1           (const unsigned char *) "The signature for Musig2 Schnorr message 1."
+#define TAG_1           (const unsigned char *) "MSG_1"
+#define MSG_1_LEN      43
+
+#define MSG_2           (const unsigned char *) "Musig2 Schnorr MSG 2"
+#define TAG_2           (const unsigned char *) "MSG_2"
+#define MSG_2_LEN      20
